pyproject.toml
[tool.poetry]
name = "picknblend"
# NOTE: This version is a placeholder!
# The version is automatically generated by poetry-dynamic-versioning
# to include the commit hash. To change the version of the package,
# create a git tag for a commit available on the main branch instead!
# For example:
#   git tag vX.Y.Z <commit-hash>
# The tag must then be pushed to the corresponding remote:
#   git push origin tag vX.Y.Z
# The resulting version will then be: X.Y.Z+HASH
# Do not tag commits in side branches! Tags do not follow commits,
# and modifying any commits in a development branch will not update
# the tag. Only create tags for commits that won't be modified
# anymore (i.e, main branch).
version = "0.0.0"
description = "PNP model placer for Blender models"
authors = ["Antmicro <contact@antmicro.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "**/*.py", from = "src/" },
    { include = "picknblend/templates/kbe_blendcfg.yaml", from = "src/" },
]

[tool.poetry.scripts]
picknblend = "picknblend.picknblend:main"

[tool.poetry.dependencies]
python = "3.11.*"
pyyaml = "^6.0.1"
wand = "^0.6.13"
bpy = "4.1"
unidecode = "^1.3.8"
hiyapyco = "^0.6.0"
pyxdg = "^0.28"

[tool.poetry-dynamic-versioning]
enable = true
metadata = true
dirty = true
ignore-untracked = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.ruff]
select = ["E", "F", "N", "B", "D", "A", "RET"]
line-length = 120
target-version = "py311"
exclude = [
	"documentation",
	"__init__.py",
]

[tool.black]
line-length = 120

[tool.mypy]
disallow_any_generics = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
